requirements.txt
#1. django 5.0.1이상 6.0.0 미만
django>=5.0.1,<6.0.0

#2. django rest framework
djangorestframework>=3.14.0,<4.0.0

# 3. psycopg2 - postgresql adapter
psycopg2>=2.9.9,<3.0.0

# 4. drf-spectacular
drf-spectacular>=0.27.1,<0.28.0

# 5. channels (socket io)
channels>=4.0.0,<4.0.1

# 6. daphne
daphne>=4.1.0,<4.2.0

# 7. uWSGI
uWSGI>=2.0.24,<2.1.0